fix: record project release event
